init project
Tapping on the plus (+) button on the top-right adds a new entry to the
list of walks. The image represents the dog you’re currently walking,
but otherwise does nothing.
The app has all the functionality it needs, except for one important
feature: The list of walks doesn’t persist. If you terminate Dog Walk
and re-launch, your entire history is gone.